Add assertion signing and name ID format
I just set up a Zammad system using Authentik SAML following these instructions. Zammad will not allow the Name Identifier Format to be left blank (which these instructions previously said to do). Zammad would not accept SAML logins until I turned on signing of assertions, so I conclude that is now mandatory. I found something online that said to remove the ---BEGIN and ---END lines from the certificate. It worked for me both ways (with and without those lines), but I wrote the documentation to say to include them. I think that's easiest and most intuitive. I tested this with Authentik 2025.4.1 and Zammad 6.4.1.

Signed-off-by: Paco Hope <pacohope@users.noreply.github.com>